requirements.txt
djangorestframework
Django==2.0.7
pytz==2018.5
